New dev section, New method of dealing with logos
The "dev" section has the sass source of the CSS files,
a makefile for same, a script to update the themes.json file.

The logo stuff is completely revamped. Before, I just had a couple of
logos which you could switch between by changing the logoid in the page,
the logo stuff was all CSS.
Now the logo is displayed via an IMG tag in the page header (with a link as
before) rather than background CSS. However, the logo is not in a static
public directory (whereas the stylesheets are), the logo is served
by the plugin itself, from the "/foil/logo" route. This enables it to be
completely customizable. The config gives a path to the desired logo
file, and the plugin will read that file and serve it when called for.
If there is no logo in the config, a default logo is used. (With the
word "FOIL"!)

I say file, but there should be two files: a dark logo and a light logo.
This is because some themes have a light header (and need a dark logo)
and some themes have a dark header (and need a light logo).
The generation of the themes.json file has been altered to include this
information about each theme.