No more redirection.
While it was cool for the theme to "magically" change by sending you
back where you came from, when there is a reverse proxy involved, it
goes wrong. There's probably some setting or argument I'm missing.
But the amount of time I spent trying to track this down, I'd rather use
the workaround of not using redirection at all, at least for the moment.
So now, when you change theme, you get shown a page which says what the
new theme is, and has a link back to where you were before.