Changed how the logo stuff is done.
Now, rather than providing a filename, the config must provide a URL for
the desired logo. Yes, that means that the caller is required to place
the logo somewhere accessible, but the caller already has to do that
with whatever else it is doing (like its own css files).
This simplifies things at this end, and also makes things faster,
because the webserver can serve a static file and know it hasn't
changed, rather than getting a dynamic response every time, where it
re-reads and re-displays the logo file on every page-load.